+++ README.md
@@ -2,7 +2,7 @@
 
 ระบบงานการเจ้าหน้าที่ตัวอย่าง พัฒนาโดยใช้ Python Flask
 
-เวอร์ชัน: 0.29.0
+เวอร์ชัน: 0.30.0
 
 เวลาตั้งต้นของการเข้างานคือ **08:30** และเวลาออกงานคือ **16:30** สามารถปรับได้ผ่านตัวแปรสภาพแวดล้อม `DEFAULT_CHECKIN_TIME` และ `DEFAULT_CHECKOUT_TIME`
 
@@ -38,7 +38,8 @@
 - `/hrm/behaviors` – หน้าบันทึกพฤติกรรม
 - `/hrm/backup` – หน้าสำรองข้อมูล
 - `/hrm/about` – หน้าเกี่ยวกับระบบ/คู่มือการใช้งาน
-
+### โครงสร้างโค้ด
+ไฟล์ route ถูกแยกออกเป็นโมดูลย่อยภายใต้โฟลเดอร์ `routes/` เพื่อให้ดูแลรักษาได้ง่ายขึ้น
 
 หากยังไม่ได้กำหนดค่า SSO สามารถทดสอบล็อกอินโดยระบุชื่อลงในพารามิเตอร์ `user` ของลิงก์ เช่น
